Make specification of includes more flexible in interface files.
Instead of allowing only include declarations of the form '#include (<FileName>|"FileName")'
in interface files, allow any text to be specified at the top of the file - including class
and typedef declarations.

Interface files are now defined to consist of a header, consisting of any text up until the
first occurence of the keyword "interface", followed by zero or more interface blocks.